# 4.x: change AMQP on disk message format & speed up the AMQP parser