--- a/deps/amqp10_common/.gitignore
+++ b/deps/amqp10_common/.gitignore
@@ -14,12 +14,12 @@
 /logs/
 /plugins/
 /plugins.lock
-/rebar.config
 /rebar.lock
 /sbin/
 /sbin.lock
 /test/ct.cover.spec
 /xrefr
+_build
 
 /amqp10_common.d
 /*.plt