--- a/deps/amqp10_common/app.bzl
+++ b/deps/amqp10_common/app.bzl
@@ -110,3 +110,13 @@ def test_suite_beam_files(name = "test_suite_beam_files"):
         app_name = "amqp10_common",
         erlc_opts = "//:test_erlc_opts",
     )
+    erlang_bytecode(
+        name = "prop_SUITE_beam_files",
+        testonly = True,
+        srcs = ["test/prop_SUITE.erl"],
+        outs = ["test/prop_SUITE.beam"],
+        hdrs = ["include/amqp10_framing.hrl"],
+        app_name = "amqp10_common",
+        erlc_opts = "//:test_erlc_opts",
+        deps = ["@proper//:erlang_app"],
+    )