--- a/deps/amqp10_common/rebar.config
+++ b/deps/amqp10_common/rebar.config
@@ -1,0 +1,5 @@
+{profiles,
+ [{test, [{deps, [proper
+                 ]}]}
+ ]
+}.